"set parent activities and actionbar with refresh button. also Askhseis, Dialekseis and Anakoinwseis are now extending MainBase for simplicity"

+++ MathLearn/app/src/main/java/com/example/marios/mathlearn/Dialekseis.java
@@ -1,5 +1,7 @@
 package com.example.marios.mathlearn;
 
+import android.app.ActionBar;
+import android.app.Dialog;
 import android.app.ProgressDialog;
 import android.content.ContentValues;
 import android.content.Context;
@@ -14,6 +16,10 @@ import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
@@ -38,54 +44,75 @@ import java.net.URL;
 import java.net.URLEncoder;
 import java.util.List;
 
-public class Dialekseis extends AppCompatActivity {
+public class Dialekseis extends MainBase {
 
     private static final String URL = String.format("http://mlearning-projectmr.rhcloud.com/videos.php");
-    private ListView listView;
-    //SQLiteDatabase db;
     Video[] videos;
-    DataBaseHelper dbHelper;
-    String codess;
+    String ids;
+    int[] idarray;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_dialekseis);
-        listView=(ListView) findViewById(R.id.listView);
-        dbHelper = new DataBaseHelper(this);
+
         videos = dbHelper.getVideos();
         StringBuilder sb= new StringBuilder();
-        String ids;
+        idarray = new int[videos.length];
         for (int i=0; i < videos.length; i++){
             sb.append( "'"+videos[i].videoID+"'," );
+            idarray[i]=videos[i].videoID;
         }
         ids=sb.toString();
-        ids = ids.substring(0, ids.length()-1);
-        new GetVideosTask(listView,ids,this).execute(URL);
+        ids = ids.substring(0, ids.length() - 1);
 
+        adapter = new CustomAdapter(Dialekseis.this);
+        adapter.setVideos(videos);
+        listView.setAdapter(adapter);
+        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                if(!videos[position].videoSelected){
+                    dbHelper.updateSelectedinVids(videos[position].videoID);
+                }
+                Intent intent = new Intent(Dialekseis.this, ProvolhDialekshs.class);
+                intent.putExtra("video", videos[position].videoLink);
+                intent.putExtra("position",position);
+                intent.putExtra("idarray",idarray);
+                intent.putExtra("title",videos[position].videoTitle);
+                startActivity(intent);
+
+            }
+        });
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.menuRefresh:
+                new GetVideosTask().execute(URL,ids);
+                return true;
+        }
+        return super.onOptionsItemSelected(item);
     }
 
     private class GetVideosTask extends AsyncTask<String, Void, String> {
-        private ListView listView;
-        private ProgressDialog dialog = new ProgressDialog(Dialekseis.this);
+        /*private ListView listView;
         private String ids;
-        private Context context;
 
-        public GetVideosTask(ListView listView, String c, Context con) {
+        public GetVideosTask(ListView listView, String c) {
             this.listView = listView;
             this.ids=c;
-            this.context=con;
-        }
+        }*/
 
         @Override
         protected void onPreExecute() {
-            this.dialog.setMessage("ΦΟΡΤΩΣΗ ΔΙΑΛΕΞΕΩΝ");
-            this.dialog.show();
+            setRefreshActionButtonState(true);
         }
 
         @Override
         protected String doInBackground(String... strings) {
-            String videos = "UNDEFINED";
+            return connectToRemote(strings[0],strings[1]);
+            /*String vids = "UNDEFINED";
             try {
                 DataOutputStream printout;
                 URL url = new URL(strings[0]);
@@ -93,12 +120,10 @@ public class Dialekseis extends AppCompatActivity {
                 urlConnection.setRequestMethod("POST");
                 urlConnection.setDoInput(true);
                 urlConnection.setDoOutput(true);
-                Uri.Builder build = new Uri.Builder()
-                        .appendQueryParameter("ids", this.ids);
+                Uri.Builder build = new Uri.Builder().appendQueryParameter("ids", this.ids);
                 String query = build.build().getEncodedQuery();
                 OutputStream os = urlConnection.getOutputStream();
-                BufferedWriter writer = new BufferedWriter(
-                        new OutputStreamWriter(os, "UTF-8"));
+                BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(os, "UTF-8"));
                 writer.write(query);
                 writer.flush();
                 writer.close();
@@ -112,80 +137,62 @@ public class Dialekseis extends AppCompatActivity {
                 while ((inputString = bufferedReader.readLine()) != null) {
                     builder.append(inputString);
                 }
-                videos = builder.toString();
+                vids = builder.toString();
 
                 urlConnection.disconnect();
-                Toast.makeText(Dialekseis.this, videos, Toast.LENGTH_LONG).show();
             } catch (Exception e) {
                 e.printStackTrace();
             }
-            return videos;
+            return vids;*/
         }
 
         @Override
         protected void onPostExecute(String temp) {
-            Toast.makeText(Dialekseis.this, temp, Toast.LENGTH_LONG).show();
             try{
                 JSONObject topLevel = new JSONObject(temp);
                 JSONArray vidjson = (JSONArray) topLevel.get("videos");
-                //StringBuilder sb= new StringBuilder();
                 String nonexistentvideos = (String) topLevel.get("nonexistentvideos");
-                //JSONObject nonexistentvideos = topLevel.getJSONObject("nonexistentvideos");
-                //String ids;
-                for (int i=0; i < vidjson.length(); i++)
-                {
-                    try {
-                        JSONObject oneObject = vidjson.getJSONObject(i);
-                        // Pulling items from the array
-                        int id = oneObject.getInt("ID");
-                        //sb.append( "'"+Integer.toString(id)+"'," );
-                        //codes[i]=code;
-                        //if (!dbHelper.existInVideos(code)){
-                        String code = oneObject.getString("Code");
+                if (vidjson.length()!=0) {
+                    for (int i = 0; i < vidjson.length(); i++) {
+                        try {
+                            JSONObject oneObject = vidjson.getJSONObject(i);
+                            // Pulling items from the array
+                            int id = oneObject.getInt("ID");
+                            String code = oneObject.getString("Code");
                             String title = oneObject.getString("Title");
                             int sequence = oneObject.getInt("Sequence");
                             dbHelper.insertInVids(new Video(title, code, false, false, sequence, id));
-                       // }
-                    } catch (JSONException e) {
-                        // Oops
+                        } catch (JSONException e) {
+                            // Oops
+                        }
                     }
                 }
-                //ids = sb.toString();
-                //ids = ids.substring(0, ids.length()-1);
-                dbHelper.deleteFromVidsIn(nonexistentvideos);
-
+                if (!nonexistentvideos.equals("")){
+                    dbHelper.deleteFromVidsIn(nonexistentvideos);
+                }
+                if (vidjson.length()==0 && nonexistentvideos.equals("") ){
+                    Toast.makeText(Dialekseis.this, "Τίποτα νεότερο.", Toast.LENGTH_LONG).show();
+                }else{
+                    videos = dbHelper.getVideos();
+                    StringBuilder sb= new StringBuilder();
+                    idarray = new int[videos.length];
+                    for (int i=0; i < videos.length; i++){
+                        sb.append( "'"+videos[i].videoID+"'," );
+                        idarray[i]=videos[i].videoID;
+                    }
+                    ids=sb.toString();
+                    ids = ids.substring(0, ids.length() - 1);
+                    //adapter = new CustomAdapter(Dialekseis.this);
+                    adapter.setVideos(videos);
+                    listView.setAdapter(adapter);
+                    Toast.makeText(Dialekseis.this, "Οι διαλέξεις ανανεώθηκαν!", Toast.LENGTH_LONG).show();
+                }
             }catch(JSONException e){
                 //No connectivity?
             }
-
-            //java.util.Arrays.toString(codes).replace('[', '(').replace(']', ')');
-
-            videos = dbHelper.getVideos();
-            if (dialog.isShowing()) {
-                dialog.dismiss();
-            }
-            CustomAdapter adapter = new CustomAdapter(Dialekseis.this);
-            adapter.setVideos(videos);
-            listView.setAdapter(adapter);
-
-            listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-                @Override
-                public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                    Intent intent = new Intent(Dialekseis.this, ProvolhDialekshs.class);
-                    intent.putExtra("video", videos[position].videoLink);
-                    intent.putExtra("id",videos[position].videoID);
-                    startActivity(intent);
-                    dbHelper.updateSelectedinVids(videos[position].videoID);
-                }
-            });
+            setRefreshActionButtonState(false);
 
         }
     }
-
-    public void goBack(View view){
-        Intent intent = new Intent(this, MainActivity.class);
-        startActivity(intent);
-    }
-
 
 }
